requirements.txt
discord.py==2.3.2
firebase-admin==6.3.0
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1